[Fix #7] Make "git fetch $AUTODOC_REMOTE" more robust
This fix handles the situation where `autodoc.sh` is running in Travis
CI, and possibly other CI environments, where only a branch of the
repo is cloned. When that is the case, you can't depend on
"refs/remotes/origin/gh-pages" to exist already. This syntax makes
sure the remote (if it exists) is retrieved.